+++ asv_bench/benchmarks/algorithms.py
@@ -3,7 +3,6 @@ from importlib import import_module
 import numpy as np
 import pandas as pd
 from pandas.util import testing as tm
-from pandas.core.algorithms import checked_add_with_arr
 
 for imp in ['pandas.util', 'pandas.tools.hashing']:
     try:
@@ -86,49 +85,6 @@ class Match(object):
 
     def time_match_string(self):
         pd.match(self.all, self.uniques)
-
-
-class AddOverflowScalar(object):
-
-    goal_time = 0.2
-
-    params = [1, -1, 0]
-    param_names = ['scalar']
-
-    def setup(self, scalar):
-        N = 10**6
-        self.arr = np.arange(N)
-
-    def time_add_overflow_scalar(self, scalar):
-        checked_add_with_arr(self.arr, scalar)
-
-
-class AddOverflowArray(object):
-
-    goal_time = 0.2
-
-    def setup(self):
-        np.random.seed(1234)
-        N = 10**6
-        self.arr = np.arange(N)
-        self.arr_rev = np.arange(-N, 0)
-        self.arr_mixed = np.array([1, -1]).repeat(N / 2)
-        self.arr_nan_1 = np.random.choice([True, False], size=N)
-        self.arr_nan_2 = np.random.choice([True, False], size=N)
-
-    def time_add_overflow_arr_rev(self):
-        checked_add_with_arr(self.arr, self.arr_rev)
-
-    def time_add_overflow_arr_mask_nan(self):
-        checked_add_with_arr(self.arr, self.arr_mixed, arr_mask=self.arr_nan_1)
-
-    def time_add_overflow_b_mask_nan(self):
-        checked_add_with_arr(self.arr, self.arr_mixed,
-                             b_mask=self.arr_nan_1)
-
-    def time_add_overflow_both_arg_nan(self):
-        checked_add_with_arr(self.arr, self.arr_mixed, arr_mask=self.arr_nan_1,
-                             b_mask=self.arr_nan_2)
 
 
 class Hashing(object):
